# src/routes/pageRoutes.js
const express = require('express');
const db = require('../database');
const { buildAioFormHtml } = require('../utils/ecpay');
const router = express.Router();

// Helper to render with front layout
function renderFront(res, page, locals = {}) {
  res.render('pages/' + page, { layout: 'front', ...locals }, function (err, body) {
    if (err) return res.status(500).send(err.message);
    res.render('layouts/front', { body, ...locals });
  });
}

// Helper to render with admin layout
function renderAdmin(res, page, locals = {}) {
  res.render('pages/admin/' + page, locals, function (err, body) {
    if (err) return res.status(500).send(err.message);
    res.render('layouts/admin', { body, ...locals });
  });
}

// Front pages
router.get('/', function (req, res) {
  renderFront(res, 'index', { title: '首頁', pageScript: 'index' });
});

router.get('/products/:id', function (req, res) {
  renderFront(res, 'product-detail', {
    title: '商品詳情',
    pageScript: 'product-detail',
    productId: req.params.id
  });
});

router.get('/cart', function (req, res) {
  renderFront(res, 'cart', { title: '購物車', pageScript: 'cart' });
});

router.get('/checkout', function (req, res) {
  renderFront(res, 'checkout', { title: '結帳', pageScript: 'checkout' });
});

router.get('/login', function (req, res) {
  renderFront(res, 'login', { title: '登入', pageScript: 'login' });
});

router.get('/orders', function (req, res) {
  renderFront(res, 'orders', { title: '我的訂單', pageScript: 'orders' });
});

router.get('/orders/:id', function (req, res) {
  renderFront(res, 'order-detail', {
    title: '訂單詳情',
    pageScript: 'order-detail',
    orderId: req.params.id,
    paymentResult: req.query.payment || ''
  });
});

// ECPay payment form page
router.get('/ecpay/payment/:orderId', function (req, res) {
  const order = db.prepare('SELECT * FROM orders WHERE id = ?').get(req.params.orderId);
  if (!order) {
    return res.status(404).send('訂單不存在');
  }
  if (order.status !== 'pending') {
    return res.redirect('/orders/' + order.id);
  }
  const items = db.prepare('SELECT product_name, product_price, quantity FROM order_items WHERE order_id = ?').all(order.id);

  // 將運費列為獨立品項，使綠界付款頁的商品明細總額與 TotalAmount 一致（拆分商品與運費）
  if (order.shipping_fee > 0) {
    const methodLabel = order.shipping_method === 'cvs' ? '超商取貨' : '宅配';
    const extras = [];
    if (order.is_remote) extras.push('偏遠地區');
    if (order.is_urgent) extras.push('當日急件');
    const shippingLabel = '運費（' + [methodLabel, ...extras].join('・') + '）';
    items.push({ product_name: shippingLabel, product_price: order.shipping_fee, quantity: 1 });
  }

  const html = buildAioFormHtml(order, items);
  res.type('text/html').send(html);
});

// Admin pages
router.get('/admin/products', function (req, res) {
  renderAdmin(res, 'products', {
    title: '商品管理',
    pageScript: 'admin-products',
    currentPath: '/admin/products'
  });
});

router.get('/admin/orders', function (req, res) {
  renderAdmin(res, 'orders', {
    title: '訂單管理',
    pageScript: 'admin-orders',
    currentPath: '/admin/orders'
  });
});

module.exports = router;